Refactor: utils 폴더 생성
